# server/src/index.ts
import express, { Application } from 'express'; //libreria
import indexRoutes from './indexRoutes';
import usuariosRoutes from './routes/usuariosRoutes';
import estudiantesRoutes from './routes/estudiantesRoutes';
import librosRoutes from './routes/librosRoutes';
import prestamosRoutes from './routes/prestamosRoutes';
import carrerasRoutes from './routes/carrerasRoutes';
import morgan from 'morgan';
import cors from 'cors';
import swagger_ui_express from 'swagger-ui-express';
import swaggerDocument from './swagger.json';
import fs from 'fs'; //libreria para guardar archivos
class Server //clase
{
    public app: Application; //variable de control
    constructor() {
        this.app = express(); //ejecutar servidor
        this.config();
        this.routes();
        this.app.use('/documentacion', swagger_ui_express.serve, swagger_ui_express.setup(swaggerDocument));
        this.app.use(express.static(__dirname + "/img"));
    }
    config(): void //definir propiedades del servidor (en este caso el puerto)
    {
        this.app.set('port', process.env.PORT || 3002);
        this.app.use(morgan('dev'));
        this.app.use(cors());
        this.app.use(express.json());
        this.app.use(express.urlencoded({ extended: false }));
    }
    routes(): void {
        this.app.use(indexRoutes);
        this.app.use('/api/usuarios', usuariosRoutes);
        this.app.use('/api/estudiantes', estudiantesRoutes);
        this.app.use('/api/libros', librosRoutes);
        this.app.use('/api/prestamos', prestamosRoutes);
        this.app.use('/api/carreras', carrerasRoutes);

        
        this.app.post('/uploadImagen', (req, res) => {
            const file = req.body.src;
            const carpeta = req.body.carpeta;
            const name = req.body.id;
            console.log(file, carpeta, name);

            const binaryData =
                Buffer.from(file.replace(/^data:image\/[a-z]+;base64,/, ""),
                    'base64').toString('binary');
            fs.writeFile(`${__dirname}/img/` + carpeta + '/' + name + '.jpg', binaryData,
                "binary", (err) => {
                console.log(err);
            });

            res.json({ fileName: name + '.jpg' });
        });




    }
    start(): void {
        this.app.listen(this.app.get('port'), () => {
            console.log('Servidor se encuentra en el puerto: ', this.app.get('port'));
        });
    }
    }
    const server = new Server();
    server.start();